Move types inside src
